Add a 'compose tar' command
Run bdcs-cli compose tar <recipe-name> to create a <recipe-name>.tar
file with all of the recipe's dependencies.

This is a bit of a hybrid at the moment -- It requires the API server,
the metadata.db, and content store repo.
You can add their paths using --mddb and --repo.

It also needs the export tool from bdcs to be available on the PATH.