Add error handling to Git Recipe code
This requires a libgit2-glib library with correct return type
annotations so that Maybe is returned in places that could return NULL.

It uses Control.Exceptions to throw a GitError when something goes
wrong. The maybeThrow helper can be used to extract the value from the
Maybe and throw the error when it is Nothing. Like this:

repo <- Git.repositoryInitRepository gfile True >>= maybeThrow CreateRepoError

repo will end up with the Git.Repository if it is successful, or a
CreateRepoError will be thrown, generating an exception if it isn't
caught.