Add message if rm fails to remove some files
dnf install requires root and some of the temporary directories
are owned by root. Not all tests are executed as root though.

Instead of using sudo and delete everything (possibly breaking
the file system in the process) we just issue a message for the
user. This also has the nice side effect of returning an exit code
of 0 so that make doesn't exit and CI does not report failures.

If temporary files are left around this is mostly problem for local
development since in CI the worker nodes are destroyed once testing
is completed.