Add some of the compose commands, simplify error handling
In the process of adding compose commands I ended up having to rework
error reporting in bdcs-api and handling here in bdcs-cli. But going
forward this should make things easier.

APIResponse is now used for all non-200 status responses, and it
includes a list of error strings. Other errors that are included with a
response are simply a [String] so that it is easier to parse.

APIResponseJSON is now APIResponse.
RecipesAPIError is now gone, APIResponse handles all the places this was
used, or a simple String list.

compose types and compose start commands have been added. Errors are
handled by printing them to the console, including unknown server
status, and problems parsing the JSON.